+++ DesignToolbox/DesignToolboxSnapshotsTests/__Snapshots__/Components/OrangeBusinessToolsTheme/Components/Indicators/OrangeBusinessToolsTheme+TagUITests.swift
@@ -41,6 +41,18 @@ final class OrangeBusinessToolsThemeTagUITests: TagUITestsTestCase {
         let interfaceStyle = UIUserInterfaceStyle.dark
         testAllTags(theme: theme, interfaceStyle: interfaceStyle)
     }
+
+    /// Tests all tags configuration in the `OrangeBusinessToolsTheme` with the `light` color scheme.
+    @MainActor func testAllInputTagsOrangeBusinessToolsThemeThemeLight() {
+        let interfaceStyle = UIUserInterfaceStyle.light
+        testAllInputTags(theme: theme, interfaceStyle: interfaceStyle)
+    }
+
+    /// Tests all tags configuration in the `OrangeBusinessToolsTheme` with the `dark` color scheme.
+    @MainActor func testAllInputTagsOrangeBusinessToolsThemeThemeDark() {
+        let interfaceStyle = UIUserInterfaceStyle.dark
+        testAllInputTags(theme: theme, interfaceStyle: interfaceStyle)
+    }
 }
 
 // swiftlint:enable required_deinit